Move module to rxjava, and change rxjava2 to rxjava
