+++ settings.gradle
@@ -7,7 +7,7 @@ includeWithFlags ':bom',                       'bom'
 
 // Published Java projects
 includeWithFlags ':core',                       'java', 'publish', 'shade', 'trim'
-includeWithFlags ':common:rxjava2',             'java', 'publish', 'relocate'
+includeWithFlags ':rxjava',                     'java', 'publish', 'relocate'
 includeWithFlags ':grpc',                       'java', 'publish', 'relocate'
 includeWithFlags ':jetty',                      'java', 'publish', 'relocate'
 includeWithFlags ':kafka',                      'java', 'publish', 'relocate'
